refactor quick fix
