+++ plugins/com.google.cloud.tools.eclipse.appengine.validation/src/com/google/cloud/tools/eclipse/appengine/validation/ApplicationQuickFix.java
@@ -16,56 +16,13 @@
 
 package com.google.cloud.tools.eclipse.appengine.validation;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.URL;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import javax.xml.transform.TransformerException;
-import org.eclipse.core.resources.IFile;
-import org.eclipse.core.resources.IMarker;
-import org.eclipse.core.runtime.CoreException;
-import org.eclipse.jface.text.IDocument;
-import org.eclipse.ui.IMarkerResolution;
-import com.google.cloud.tools.eclipse.util.Xslt;
-
 /**
  * Applies application.xsl to appengine-web.xml to remove an <application/> element.
  */
-public class ApplicationQuickFix implements IMarkerResolution {
+public class ApplicationQuickFix extends XsltQuickFix {
   
-  private static final Logger logger = Logger.getLogger(
-      ApplicationQuickFix.class.getName());
-
-  @Override
-  public String getLabel() {
-    return Messages.getString("remove.application.element");
-  }
-  
-  /**
-   * Attempts to edit the {@link IDocument} in the open editor. If the editor is not open,
-   * reads the file from memory and transforms in place.
-   */
-  @Override
-  public void run(IMarker marker) {
-    try {
-      IFile file = (IFile) marker.getResource();
-      IDocument document = ValidationUtils.getCurrentDocument(file);
-      URL xslPath = ApplicationQuickFix.class.getResource("/xslt/application.xsl");
-      if (document != null) {
-        String currentContents = document.get();
-        String encoding = file.getCharset();
-        InputStream documentStream = new ByteArrayInputStream(currentContents.getBytes(encoding));
-        InputStream transformed = Xslt.applyXslt(documentStream, xslPath.openStream());
-        String newDoc = ValidationUtils.convertStreamToString(transformed, encoding);
-        document.set(newDoc);
-      } else {
-        Xslt.transformInPlace(file, xslPath);
-      }
-    } catch (IOException | TransformerException | CoreException ex) {
-      logger.log(Level.SEVERE, ex.getMessage());
-    }
+  public ApplicationQuickFix() {
+    super("/xslt/application.xsl", Messages.getString("remove.application.element"));
   }
   
 }
