+++ plugins/com.google.cloud.tools.eclipse.appengine.validation/src/com/google/cloud/tools/eclipse/appengine/validation/ToServlet25QuickFix.java
@@ -16,56 +16,14 @@
 
 package com.google.cloud.tools.eclipse.appengine.validation;
 
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.URL;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import javax.xml.transform.TransformerException;
-import org.eclipse.core.resources.IFile;
-import org.eclipse.core.resources.IMarker;
-import org.eclipse.core.runtime.CoreException;
-import org.eclipse.jface.text.IDocument;
-import org.eclipse.ui.IMarkerResolution;
-import com.google.cloud.tools.eclipse.util.Xslt;
-
 /**
- * Replaces web.xml <web-app> element with Servlet 2.5 deployment descriptor.
+ * Applies servlet.xsl to web.xml to replace <web-app> element with
+ * Servlet 2.5 deployment descriptor.
  */
-public class ToServlet25QuickFix implements IMarkerResolution {
+public class ToServlet25QuickFix extends XsltQuickFix {
   
-  private static final Logger logger = Logger.getLogger(
-      ToServlet25QuickFix.class.getName());
-
-  @Override
-  public String getLabel() {
-    return Messages.getString("convert.servlet");
-  }
-  
-  /**
-   * Attempts to edit the {@link IDocument} in the open editor. If the editor is not open,
-   * reads the file from memory and transforms in place.
-   */
-  @Override
-  public void run(IMarker marker) {
-    try {
-      IFile file = (IFile) marker.getResource();
-      IDocument document = ValidationUtils.getCurrentDocument(file);
-      URL xslPath = ApplicationQuickFix.class.getResource("/xslt/servlet.xsl");
-      if (document != null) {
-        String currentContents = document.get();
-        String encoding = file.getCharset();
-        InputStream documentStream = new ByteArrayInputStream(currentContents.getBytes(encoding));
-        InputStream transformed = Xslt.applyXslt(documentStream, xslPath.openStream());
-        String newDoc = ValidationUtils.convertStreamToString(transformed, encoding);
-        document.set(newDoc);
-      } else {
-        Xslt.transformInPlace(file, xslPath);
-      }
-    } catch (IOException | TransformerException | CoreException ex) {
-      logger.log(Level.SEVERE, ex.getMessage());
-    }
+  public ToServlet25QuickFix() {
+    super("/xslt/servlet.xsl", Messages.getString("convert.servlet"));
   }
   
 }
